+++ components/faphub/appcard/composable/src/main/java/com/flipperdevices/faphub/appcard/composable/paging/ComposableFapsList.kt
@@ -14,11 +14,12 @@ import androidx.paging.LoadState
 import androidx.paging.compose.LazyPagingItems
 import androidx.paging.compose.itemContentType
 import androidx.paging.compose.itemKey
-import com.flipperdevices.core.ui.errors.ComposableThrowableError
 import com.flipperdevices.core.ui.ktx.image.ComposeLottiePic
 import com.flipperdevices.core.ui.theme.LocalPallet
 import com.flipperdevices.faphub.appcard.composable.AppCard
 import com.flipperdevices.faphub.dao.api.model.FapItemShort
+import com.flipperdevices.faphub.errors.api.FapErrorSize
+import com.flipperdevices.faphub.errors.api.FapHubComposableErrorsRenderer
 import com.flipperdevices.core.ui.res.R as DesignSystem
 
 private const val DEFAULT_FAP_COUNT = 20
@@ -27,6 +28,8 @@ private const val DEFAULT_FAP_COUNT = 20
 fun LazyListScope.ComposableFapsList(
     faps: LazyPagingItems<FapItemShort>,
     onOpenFapItem: (FapItemShort) -> Unit,
+    errorsRenderer: FapHubComposableErrorsRenderer,
+    defaultFapErrorSize: FapErrorSize,
     installationButton: @Composable (FapItemShort?, Modifier) -> Unit
 ) {
     val elementModifier = Modifier
@@ -45,11 +48,12 @@ fun LazyListScope.ComposableFapsList(
         return
     } else if (faps.loadState.refresh is LoadState.Error) {
         val loadState = faps.loadState.refresh as? LoadState.Error ?: return
-        item {
-            ComposableThrowableError(
+        with(errorsRenderer) {
+            ComposableThrowableErrorListItem(
                 modifier = elementModifier,
                 throwable = loadState.error,
-                onRetry = faps::retry
+                onRetry = faps::retry,
+                fapErrorSize = defaultFapErrorSize
             )
         }
         return
@@ -73,11 +77,12 @@ fun LazyListScope.ComposableFapsList(
                 ComposableLoadingItem()
             }
 
-            is LoadState.Error -> item {
-                ComposableThrowableError(
+            is LoadState.Error -> with(errorsRenderer) {
+                ComposableThrowableErrorListItem(
                     throwable = loadState.error,
                     onRetry = faps::retry,
-                    modifier = elementModifier
+                    modifier = elementModifier,
+                    fapErrorSize = FapErrorSize.IN_LIST
                 )
             }