+++ components/faphub/dao/network/src/main/java/com/flipperdevices/faphub/dao/network/ktorfit/model/KtorfitExceptionDetail.kt
@@ -1,0 +1,14 @@
+package com.flipperdevices.faphub.dao.network.ktorfit.model
+
+import kotlinx.serialization.SerialName
+import kotlinx.serialization.Serializable
+
+@Serializable
+data class KtorfitExceptionDetail(
+    @SerialName("status")
+    val status: String,
+    @SerialName("code")
+    val code: Int,
+    @SerialName("details")
+    val details: String
+)