+++ components/faphub/errors/api/src/main/java/com/flipperdevices/faphub/errors/api/throwable/FirmwareNotSupported.kt
@@ -1,0 +1,3 @@
+package com.flipperdevices.faphub.errors.api.throwable
+
+class FirmwareNotSupported(origin: Throwable? = null) : Throwable(origin)