+++ components/faphub/installation/manifest/impl/src/main/java/com/flipperdevices/faphub/installation/manifest/impl/utils/FapManifestsLoader.kt
@@ -9,7 +9,7 @@ import com.flipperdevices.core.ktx.jre.flatten
 import com.flipperdevices.core.ktx.jre.pmap
 import com.flipperdevices.core.log.LogTagProvider
 import com.flipperdevices.core.log.info
-import com.flipperdevices.core.ui.errors.throwable.FlipperNotConnected
+import com.flipperdevices.faphub.errors.api.throwable.FlipperNotConnected
 import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestConstants.FAP_MANIFESTS_FOLDER_ON_FLIPPER
 import com.flipperdevices.faphub.installation.manifest.impl.utils.FapManifestConstants.FAP_MANIFEST_EXTENSION
 import com.flipperdevices.faphub.installation.manifest.model.FapManifestItem