+++ components/hub/impl/src/main/java/com/flipperdevices/hub/impl/api/HubFeatureEntryImpl.kt
@@ -1,8 +1,6 @@
 package com.flipperdevices.hub.impl.api
 
 import androidx.compose.foundation.layout.padding
-import androidx.compose.runtime.collectAsState
-import androidx.compose.runtime.getValue
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import androidx.navigation.NavGraphBuilder
@@ -19,12 +17,10 @@ import com.flipperdevices.faphub.main.api.FapHubMainScreenApi
 import com.flipperdevices.faphub.maincard.api.MainCardApi
 import com.flipperdevices.hub.api.HubFeatureEntry
 import com.flipperdevices.hub.impl.composable.ComposableHub
-import com.flipperdevices.hub.impl.viewmodel.HubViewModel
 import com.flipperdevices.nfc.attack.api.NFCAttackFeatureEntry
 import com.flipperdevices.screenstreaming.api.ScreenStreamingFeatureEntry
 import com.squareup.anvil.annotations.ContributesBinding
 import com.squareup.anvil.annotations.ContributesMultibinding
-import tangle.viewmodel.compose.tangleViewModel
 import javax.inject.Inject
 
 @ContributesBinding(AppGraph::class, HubFeatureEntry::class)
@@ -54,26 +50,21 @@ class HubFeatureEntryImpl @Inject constructor(
                 arguments = hubArguments
             ) {
                 val globalNavController = LocalGlobalNavigationNavStack.current
-                val hubViewModel = tangleViewModel<HubViewModel>()
-                val isApplicationCatalogEnabled by hubViewModel.isApplicationCatalogEnabledFlow()
-                    .collectAsState()
                 ComposableHub(
                     onOpenAttack = {
                         navController.navigate(nfcAttackFeatureEntry.ROUTE.name)
                     },
                     mainCardComposable = {
-                        if (isApplicationCatalogEnabled) {
-                            mainCardApi.ComposableMainCard(
-                                modifier = Modifier.padding(
-                                    start = 14.dp,
-                                    end = 14.dp,
-                                    top = 14.dp
-                                ),
-                                onClick = {
-                                    navController.navigate(fapHubMainScreenApi.ROUTE.name)
-                                }
-                            )
-                        }
+                        mainCardApi.ComposableMainCard(
+                            modifier = Modifier.padding(
+                                start = 14.dp,
+                                end = 14.dp,
+                                top = 14.dp
+                            ),
+                            onClick = {
+                                navController.navigate(fapHubMainScreenApi.ROUTE.name)
+                            }
+                        )
                     },
                     onOpenRemoteControl = {
                         globalNavController.navigate(screenStreamingFeatureEntry.ROUTE.name)